Spread swift class
