+++ AR/Persistence/PersistenceViewController.swift
@@ -16,8 +16,6 @@ class PersistenceViewController: UIViewController , ARSCNViewDelegate , ARSessio
     @IBOutlet weak var imageView: UIImageView!
     @IBOutlet weak var statusLabel: UILabel!
     
-    var worldMappingStatus: ARFrame.WorldMappingStatus?
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         setNavigationButton()
@@ -94,141 +92,3 @@ class PersistenceViewController: UIViewController , ARSCNViewDelegate , ARSessio
     
 }
 
-
-//MARK: Setting View
-extension PersistenceViewController {
-    
-    func setNavigationButton() {
-        let saveRightBarButtonItem = UIBarButtonItem.init(title: "save", style: .done, target: self, action: #selector(PersistenceViewController.saveData))
-        
-        let loadRightBarButtonItem = UIBarButtonItem.init(title: "load", style: .done, target: self, action: #selector(PersistenceViewController.loadData))
-
-        self.navigationItem.rightBarButtonItems = [saveRightBarButtonItem, loadRightBarButtonItem]
-    }
-    
-    var mapDataFromFile: Data? {
-        return try? Data(contentsOf: mapSaveURL)
-    }
-    
-    // tap gesture
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        
-        if let touchLocation = touches.first?.location(in: SCNView) {
-            
-            let hitTestResult = SCNView.hitTest(touchLocation, types: .featurePoint)
-            
-            if let hitResult = hitTestResult.first {
-                updateText(text: "혜리", at: hitResult)
-//                putShip(at: hitResult)
-            }
-        }
-    }
-    
-    func sessionShouldAttemptRelocalization(_ session: ARSession) -> Bool {
-        return true
-    }
-    
-    func putShip (at: ARHitTestResult) {
-        
-        // Remove exisitng anchor and add new anchor
-        if let existingAnchor = virtualObjectAnchor {
-            SCNView.session.remove(anchor: existingAnchor)
-        }
-        virtualObjectAnchor = ARAnchor(name: virtualObjectAnchorName, transform: at.worldTransform)
-        SCNView.session.add(anchor: virtualObjectAnchor!)
-
-        let scene = SCNScene(named: "art.scnassets/ship.scn")!
-        SCNView.scene = scene
-
-    }
-    
-    func updateText(text: String, at: ARHitTestResult) {
-        
-        // Remove exisitng anchor and add new anchor
-        if let existingAnchor = virtualObjectAnchor {
-            SCNView.session.remove(anchor: existingAnchor)
-        }
-        
-        virtualObjectAnchor = ARAnchor(name: virtualObjectAnchorName, transform: at.worldTransform)
-        SCNView.session.add(anchor: virtualObjectAnchor!)
-        
-        let textGeometry = SCNText(string: "\(text)", extrusionDepth: 1.0)
-        textGeometry.firstMaterial?.diffuse.contents = UIColor.black
-        let textNode = SCNNode(geometry: textGeometry)
-        textNode.position = SCNVector3(at.worldTransform.columns.3.x , at.worldTransform.columns.3.y, at.worldTransform.columns.3.z)
-        textNode.scale = SCNVector3(0.01, 0.01, 0.01)
-        SCNView.scene.rootNode.addChildNode(textNode)
-        
-    }
-
-    // indicates whether it’s currently a good time to capture a world map
-    func session(_ session: ARSession, didUpdate frame: ARFrame) {
-        // Enable Save button only when the mapping status is good and an object has been placed
-        switch frame.worldMappingStatus {
-        case .extending, .mapped:
-            print("좋음")
-        default:
-            print("구림")
-        }
-        statusLabel.text = """
-        Mapping: \(frame.worldMappingStatus.description)
-        Tracking: \(frame.camera.trackingState.description)
-        """
-    }
-
-}
-
-//MARK: Persistence
-extension PersistenceViewController {
-    @objc func saveData() {
-        
-        print("save Data")
-        SCNView.session.getCurrentWorldMap { worldMap, error in
-            guard let map = worldMap else { return }
-            
-            // Add a snapshot image indicating where the map was captured.
-            guard let snapshotAnchor = SnapshotAnchor(capturing: self.SCNView)
-                else { fatalError("Can't take snapshot") }
-            map.anchors.append(snapshotAnchor)
-            
-            do {
-                let data = try NSKeyedArchiver.archivedData(withRootObject: map, requiringSecureCoding: true)
-                try data.write(to: self.mapSaveURL, options: [.atomic])
-                DispatchQueue.main.async {
-                    
-                }
-            } catch {
-                fatalError("Can't save map: \(error.localizedDescription)")
-            }
-        }
-    }
-    
-    @objc func loadData() {
-        print("load Data")
-        /// - Tag: ReadWorldMap
-        let worldMap: ARWorldMap = {
-            guard let data = mapDataFromFile
-                else { fatalError("Map data should already be verified to exist before Load button is enabled.") }
-            do {
-                guard let worldMap = try NSKeyedUnarchiver.unarchivedObject(ofClass: ARWorldMap.self, from: data)
-                    else { fatalError("No ARWorldMap in archive.") }
-                return worldMap
-            } catch {
-                fatalError("Can't unarchive ARWorldMap from file data: \(error)")
-            }
-        }()
-        
-        // Display the snapshot image stored in the world map to aid user in relocalizing.
-        if let snapshotData = worldMap.snapshotAnchor?.imageData,
-            let snapshot = UIImage(data: snapshotData) {
-            self.imageView.image = snapshot
-
-            //snapshot image 뜨게
-        } else {
-            print("No snapshot image in world map")
-        }
-        // Remove the snapshot anchor from the world map since we do not need it in the scene.
-        worldMap.anchors.removeAll(where: { $0 is SnapshotAnchor })
-
-    }
-}
